lf polish: lfd: remove backwards compatibility shims for loop-to-job rename
