--- src/loopflow/lfd/loop_runner.py
@@ -1,101 +1,0 @@
-"""Backwards compatibility - use job_runner.py instead."""
-
-# Re-export everything from job_runner.py
-from loopflow.lfd.job_runner import (
-    SCHEDULER_POLL_INTERVAL,
-    SOCKET_PATH,
-    Loop,
-    LoopRun,
-    LoopStatus,
-    _auto_merge_pr,
-    _build_job_inline_prompt,
-    _build_job_prompt,
-    _build_loop_inline_prompt,
-    _build_loop_prompt,
-    _cleanup_variant_worktrees,
-    _cleanup_worktree,
-    _create_pr_to_job_main,
-    _create_pr_to_loop_main,
-    _iteration_branch_prefix,
-    _land_to_main,
-    _run_collector_step,
-    _run_fork_join_group,
-    _scheduler_acquire,
-    _scheduler_call,
-    _scheduler_release,
-    _VariantResult,
-    count_outstanding,
-    create_worktree,
-    get_job,
-    get_loop,
-    main,
-    notify_event,
-    run_iteration,
-    run_job_iterations,
-    run_loop_iterations,
-    save_job_run,
-    save_loop_run,
-    update_job_iteration,
-    update_job_pid,
-    update_job_run_pr,
-    update_job_run_status,
-    update_job_run_step,
-    update_job_status,
-    update_loop_iteration,
-    update_loop_pid,
-    update_loop_run_pr,
-    update_loop_run_status,
-    update_loop_run_step,
-    update_loop_status,
-)
-
-__all__ = [
-    "SCHEDULER_POLL_INTERVAL",
-    "SOCKET_PATH",
-    "Loop",
-    "LoopRun",
-    "LoopStatus",
-    "_VariantResult",
-    "_auto_merge_pr",
-    "_build_job_inline_prompt",
-    "_build_job_prompt",
-    "_build_loop_inline_prompt",
-    "_build_loop_prompt",
-    "_cleanup_variant_worktrees",
-    "_cleanup_worktree",
-    "_create_pr_to_job_main",
-    "_create_pr_to_loop_main",
-    "_iteration_branch_prefix",
-    "_land_to_main",
-    "_run_collector_step",
-    "_run_fork_join_group",
-    "_scheduler_acquire",
-    "_scheduler_call",
-    "_scheduler_release",
-    "count_outstanding",
-    "create_worktree",
-    "get_job",
-    "get_loop",
-    "main",
-    "notify_event",
-    "run_iteration",
-    "run_job_iterations",
-    "run_loop_iterations",
-    "save_job_run",
-    "save_loop_run",
-    "update_job_iteration",
-    "update_job_pid",
-    "update_job_run_pr",
-    "update_job_run_status",
-    "update_job_run_step",
-    "update_job_status",
-    "update_loop_iteration",
-    "update_loop_pid",
-    "update_loop_run_pr",
-    "update_loop_run_status",
-    "update_loop_run_step",
-    "update_loop_status",
-]
-
-if __name__ == "__main__":
-    main()
