--- src/loopflow/lfd/loops.py
@@ -1,56 +1,0 @@
-"""Backwards compatibility - use jobs.py instead."""
-
-# Re-export everything from jobs.py
-from loopflow.lfd.jobs import (
-    MAGICAL,
-    MUSICAL,
-    StartResult,
-    _allocate_job_main,
-    _allocate_loop_main,
-    _branch_exists,
-    _create_job_main_branch,
-    _create_loop_main_branch,
-    _generate_random_words,
-    _run_job,
-    _run_loop,
-    count_outstanding,
-    create_job,
-    create_loop,
-    get_job,
-    get_loop,
-    get_wt_from_cwd,
-    is_process_running,
-    start_job,
-    start_loop,
-    stop_job,
-    stop_loop,
-    update_job_status,
-    update_loop_status,
-)
-
-__all__ = [
-    "MAGICAL",
-    "MUSICAL",
-    "StartResult",
-    "_allocate_job_main",
-    "_allocate_loop_main",
-    "_branch_exists",
-    "_create_job_main_branch",
-    "_create_loop_main_branch",
-    "_generate_random_words",
-    "_run_job",
-    "_run_loop",
-    "count_outstanding",
-    "create_job",
-    "create_loop",
-    "get_job",
-    "get_loop",
-    "get_wt_from_cwd",
-    "is_process_running",
-    "start_job",
-    "start_loop",
-    "stop_job",
-    "stop_loop",
-    "update_job_status",
-    "update_loop_status",
-]
